% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/california_entities.R
\name{california_lenders}
\alias{california_lenders}
\title{California Lenders Licensees}
\usage{
california_lenders(search_names = c("137", "Marble Bridge"),
  return_message = TRUE)
}
\arguments{
\item{return_message}{}
}
\description{
California Lenders Licensees
}
