% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/california_entities.R
\name{get_data_lenders_california}
\alias{get_data_lenders_california}
\title{California Lenders Licensees}
\usage{
get_data_lenders_california(search_names = c("137", "Marble Bridge"),
  return_message = TRUE)
}
\arguments{
\item{return_message}{}
}
\description{
California Lenders Licensees
}
